Estruturación do proxecto: nomes de paquetes en maiúscula #26
