+++ src/main/java/contenido/Anuncios.java
@@ -3,9 +3,9 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package Contenido;
+package contenido;
 
-import Utils.Exceptions.TitleNotFoundException;
+import utils.Exceptions.TitleNotFoundException;
 import java.util.ArrayList;
 import java.util.List;
