+++ src/main/java/servidor/ServLocal.java
@@ -3,14 +3,14 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package Servidor;
+package servidor;
 
-import Contenido.Anuncios;
-import Servidor.Tokens.Token;
-import Contenido.Contenido;
-import Servidor.Tokens.AdminToken;
-import Utils.Exceptions.ContentNotFoundException;
-import Utils.Exceptions.InvalidTokenException;
+import contenido.Anuncios;
+import servidor.Tokens.Token;
+import contenido.Contenido;
+import servidor.Tokens.AdminToken;
+import utils.Exceptions.ContentNotFoundException;
+import utils.Exceptions.InvalidTokenException;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Random;
