+++ src/main/java/servidor/Tokens/AdminToken.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package Servidor.Tokens;
+package servidor.Tokens;
 
 import java.util.Objects;
 import java.util.UUID;
