+++ src/main/java/utils/Exceptions/TitleNotFoundException.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package Utils.Exceptions;
+package utils.Exceptions;
 
 /**
  *
