+++ src/test/java/contenido/AnunciosTest.java
@@ -3,9 +3,11 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package Contenido;
+package contenido;
 
-import Utils.Exceptions.TitleNotFoundException;
+import contenido.Contenido;
+import contenido.Anuncios;
+import utils.Exceptions.TitleNotFoundException;
 import java.util.ArrayList;
 import java.util.List;
 import org.junit.After;
